[test] Expression 값 검증 분리
