+++ src/main/java/com/jaeyeonling/calculate/ExpressionValidator.java
@@ -1,0 +1,10 @@
+package com.jaeyeonling.calculate;
+
+public final class ExpressionValidator {
+
+    private ExpressionValidator() { }
+
+    public static void validate(String expression) {
+
+    }
+}
